+++ createMark.js
@@ -1,0 +1,21 @@
+
+function createReloadButton() {
+    const $wrapper = createElement('div', 'reloadWrap');
+    const $btnReload = createElement('button', 'button');
+    $btnReload.textContent = ' Restart';
+    $wrapper.appendChild($btnReload);
+    $btnReload.addEventListener('click', () => {
+        window.location.reload();
+    });
+    return $wrapper;
+}
+
+function createElement(tag, className) {
+    const $tag = document.createElement(tag);
+    if (className) {
+        $tag.classList.add(className);
+    }
+    return $tag;
+}
+
+export {createElement, createReloadButton};
