+++ getRandom.js
@@ -1,0 +1,4 @@
+export default
+function getRandom(num) {
+    return Math.ceil(Math.random() * num);
+}
