+++ render.js
@@ -1,0 +1,17 @@
+function changeHp(num) {
+    this.hp -= num;
+    if (this.hp <= 0) {
+        this.hp = 0;
+    }
+    return this.hp;
+}
+
+function elHP() {
+    return document.querySelector('.player' + this.player + ' .life');
+}
+
+function renderHP() {
+    this.elHP().style.width = this.hp + '%';
+}
+
+export { changeHp, elHP, renderHP };
